Moved PremisLogger back to events package, and brought the other event clients with it
