+++ fcrepo-clients/src/main/java/edu/unc/lib/dl/event/PremisLogger.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.util;
+package edu.unc.lib.dl.event;
 
 import java.util.Date;
 import java.util.List;
